pyproject.toml
[tool.poetry]
name = "refbooks_keeper"
version = "0.1.0"
description = ""
authors = ["Andrey_Fominykh <meriadok2003@mail.ru>"]

[tool.poetry.dependencies]
python = "3.10^"
Django = "^4.1.2"
python-dotenv = "^0.21.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.2.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
